2.15

+++ src/main/java/lang/print/gaps/finalModuleTask/TemperatureConverter.java
@@ -2,7 +2,7 @@ package lang.print.gaps.finalModuleTask;
 
 public class TemperatureConverter {
     public void toFahrenheit(int temperatureCelsius){
-        float temperaturFahrenheit = temperatureCelsius*9/5+32;
+        float temperaturFahrenheit = ((float)temperatureCelsius*9/5)+32;
         System.out.println(temperaturFahrenheit);
 
     }
